functions/requirements.txt
firebase-functions==0.1.2
firebase-admin==6.5.0
functions-framework==3.8.2
python-dotenv==1.1.0
openai==1.76.0
PyPDF2==3.0.1
langchain==0.2.17
langchain-core==0.2.43
langchain-text-splitters==0.2.4
google-cloud-storage~=2.14.0
google-cloud-firestore~=2.15.0